1st-Assignment-of-Session-3
Answer .C

This will run, print a message and terminate gracefully.Because the
runtime system needs to load the class before it can look for the main
method. So the static initializer will run first and print "10".
Immediately after that System.exit(0) will  be called terminating the
program before an error can be thrown.